+++ SUCC/DataFiles/Memory Files/MemoryReadOnlyDataFile.cs
@@ -1,5 +1,6 @@
 ﻿using SUCC.Abstractions;
 using System.IO;
+using System.Linq;
 
 namespace SUCC.MemoryFiles
 {
@@ -13,18 +14,29 @@ namespace SUCC.MemoryFiles
         /// </summary>
         /// <remarks> why would you do this? </remarks>
         public MemoryReadOnlyDataFile() : this(string.Empty) 
-        { 
+        {
         }
 
         /// <summary>
         /// Creates a ReadOnlyDataFile in memory with some preexisting SUCC content.
         /// </summary>
-        public MemoryReadOnlyDataFile(string rawFileText, string defaultFileText = null) : base(defaultFileText)
+        public MemoryReadOnlyDataFile(string rawFileText) : this(rawFileText, rawFileText)
+        {
+        }
+
+        /// <summary>
+        /// Advanced constructor, with an identifier for the file and default file text.
+        /// </summary>
+        public MemoryReadOnlyDataFile(string rawFileText, string identifier, string defaultFileText = null) : base(defaultFileText)
         {
             MemoryTextData = rawFileText;
+            Identifier = identifier;
             this.ReloadAllData();
         }
 
+
+        /// <inheritdoc/>
+        public override string Identifier { get; }
 
         private readonly string MemoryTextData = "";