+++ Changelog.md
@@ -12,6 +12,8 @@
   * todo: document this, add non-generic version, add path based versions
 * added ReadableWritableDataFile.ResetToDefaultData()
 * fixed default data path on Mac OS (Unity version)
+* added `ReadableDataFile.Identifier`
+  * when creating a `MemoryDataFile`, you can now supply an identifier
 
 todo update the docs for 1.2